[REFACTOR] - Change the server and client API to help future proof as well as clean up some edge cases
