--- src/main/java/com/xjeffrose/xio2/client/LoadBalancingStrategy.java
@@ -1,7 +1,0 @@
-package com.xjeffrose.xio2.client;
-
-import java.net.InetSocketAddress;
-
-public interface LoadBalancingStrategy {
-  InetSocketAddress nextAddress();
-}
