+++ src/main/java/com/xjeffrose/xio2/http/client/LoadBalancingStrategy.java
@@ -1,0 +1,7 @@
+package com.xjeffrose.xio2.http.client;
+
+import java.net.InetSocketAddress;
+
+public interface LoadBalancingStrategy {
+  InetSocketAddress nextAddress();
+}
