+++ src/main/java/com/xjeffrose/xio2/http/client/NullLoadBalancer.java
@@ -1,0 +1,14 @@
+package com.xjeffrose.xio2.http.client;
+
+import java.net.InetSocketAddress;
+
+public class NullLoadBalancer implements LoadBalancingStrategy {
+  private final InetSocketAddress addr;
+  public NullLoadBalancer(InetSocketAddress addr) {
+    this.addr = addr;
+  }
+  @Override
+  public InetSocketAddress nextAddress() {
+    return addr;
+  }
+}
