+++ src/main/java/com/xjeffrose/xio2/http2/Http2.java
@@ -1,0 +1,4 @@
+package com.xjeffrose.xio2.http2;
+
+public class Http2 {
+}
