--- src/main/java/com/xjeffrose/xio2/server/Handler.java
@@ -1,7 +1,0 @@
-package com.xjeffrose.xio2.server;
-
-import com.xjeffrose.xio2.http.HttpObject;
-
-public interface Handler {
-  void handle(ChannelContext ctx, HttpObject req);
-}
