+++ src/test/java/com/xjeffrose/xio2/http/server/TLS/TLSTest.java
@@ -1,0 +1,51 @@
+package com.xjeffrose.xio2.http.server.TLS;
+
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
+
+public class TLSTest {
+
+  @Before
+  public void setUp() throws Exception {
+
+  }
+
+  @After
+  public void tearDown() throws Exception {
+
+  }
+
+  @Test
+  public void testEncrypt() throws Exception {
+
+  }
+
+  @Test
+  public void testDecrypt() throws Exception {
+
+  }
+
+  @Test
+  public void testClose() throws Exception {
+
+  }
+
+  @Test
+  public void testTerminate() throws Exception {
+
+  }
+
+  @Test
+  public void testEngineBuilder() throws Exception {
+//    SSLEngine engine = TLS.engineBuilder();
+//    assertTrue(engine != null);
+//    assertEquals(engine.getEnabledProtocols()[0], "TLSv1.2");
+//    assertEquals(engine.getEnabledCipherSuites()[0], "TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256");
+  }
+
+  @Test
+  public void testEncryption() throws Exception {
+  }
+
+}
